pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "severino"
version = "0.1.0"
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "A CLI for ML monitoring assistance using Gemma."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other/Proprietary License", # Custom Non-Commercial License
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Utilities",
]
dependencies = [
    "click",
    "llama-cpp-python",
    "python-dotenv",
    "psutil",
    "rich",
    "flask", # Added for API functionality
    "faiss-cpu", # For efficient vector search
    "sentence-transformers", # For generating embeddings
]

[project.optional-dependencies]
cuda = ["llama-cpp-python[cuda]"]
dev = ["pytest"]

[project.urls]
"Homepage" = "https://github.com/your-repo/severino" # Replace with your actual repo
"Bug Tracker" = "https://github.com/your-repo/severino/issues" # Replace with your actual repo

[project.scripts]
severino = "src.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
